chore: delete main.cpp file
