+++ boxify.cpp
@@ -66,3 +66,15 @@ void Boxifier::boxify(CodeBlock& block) {
         } 
     }
 }
+
+int main() {
+    ExtensionFinder finder(EXTENSIONS);
+    
+    
+    for (std::string path : finder.getFileNames()) {
+        Boxifier boxifier(path);
+        boxifier.boxifyAll();
+    }
+
+    return 0; 
+}
